refactor: FrameHeaderField 추출
